+++ examples/big-test.ts
@@ -1,34 +1,26 @@
 // This test requires a minecraft server started with rcon
-// enabled on port 25575 and the following password
-
-const password = "password"
+// enabled on port 25575 and 'password' as password
 
 import {Rcon} from "../lib/rcon"
 
 const connectOptions = {
-  host: "localhost", port: 25575, password
+  host: "localhost", port: 25575, password: "password"
 }
 
-const rcon = new Rcon({packetResponseTimeout: 500, resendPacketOnTimeout: true})
+const rcon = new Rcon({packetResponseTimeout: 500})
 
 rcon.connect(connectOptions)
 .catch(e => console.error("Couldn't connect:", e))
 .then(() => {
-  let promises = []
+  const promises = new Array(2000).fill(null).map((_, i) => {
+    return rcon.send("say " + i)
+  })
 
-  for (let i = 0; i < 1024; i++) {
-    // promises.push(rcon.send(`help ${i % 9 + 1}`))
-    promises.push(rcon.send("say " + i))
-  }
-
-  console.time("help pages")
-  Promise.all(promises)
-  .then(helpPages => {
-    console.timeEnd("help pages")
-    console.log("===== All help pages =====")
-    console.log(helpPages[0])
-    console.log(helpPages.length)
-    rcon.disconnect()
+  console.time("promises")
+  Promise.all(promises).then(responses => {
+    console.timeEnd("promises")
+    console.log(responses.length)
+    return rcon.end()
   })
   .catch(e => console.error("Error while sending packets:", e))
 })