+++ typings/spec.d.ts
@@ -1,2 +1,0 @@
-/// <reference path="./index.d.ts"/>
-/// <reference path="globals/jasmine/index.d.ts" />