+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.1.1 (2018-09-18)
+
+- Uses async function
+- Added `static connect()` function
+
 ## 3.0.0 (2016-10-20)
 
 - Major API Changes.